# Replace the pathfinding algorithm with a faster one