--- a/Cargo.toml
+++ b/Cargo.toml
@@ -21,7 +21,6 @@ doc = true
 
 [dependencies]
 byteorder = "1.2"
-pathfinding = "1.1"
 hashbrown = "0.5"
 memmap = "0.7"