@wareya 2019-08-23
PS: notmecab already disagrees with mecab between optimal same-cost paths
pretty often, and somewhere, the readme (or documentation) of notmecab
specifies that tiebreaking is not defined, so it shouldn't be a problem.
Maybe it should specify that it's *arbitrary* but consistent within a
single version and environment (dictionaries, config, etc) instead.

On Fri, Aug 23, 2019, 15:00 Alexander Nadeau <wareya@gmail.com> wrote:

> I'm away from my development computer for a few days, I'll test and review
> this once I'm back. Thanks.
>
> On Fri, Aug 23, 2019, 14:59 Koute <notifications@github.com> wrote:
>
>> On my totally unscientific benchmarks I'm getting the following numbers:
>>
>> Before:
>>
>>    - One large string: 1010ms
>>    - Thousands of small strings: 32958ms
>>
>> After:
>>
>>    - One large string: 836ms
>>    - Thousands of small strings: 7099ms
>>
>> Your mileage may of course vary. (:
>>
>> I *think* the algorithm is correct (I've copied it from another of my
>> projects, which is why it's kinda generic.), but please try the code from
>> this PR on your own stuff before merging! I've split this change into two
>> commits - in the first part I've added the new algorithm alongside the old
>> one (so *both* are used) and I've added some asserts to verify that
>> their output matches (well, kind of - it only checks that the costs match
>> as I was too lazy to add stable disambiguation between multiple same-cost
>> paths, so it doesn't check that the output matches *exactly*). So when
>> testing this just temporarily revert the last commit.
>>
>> The new algorithm basically exploits the fact that we can easily and
>> cheaply generate a sorted list of tokens according to their position in the
>> text, and that the number of unique tokens starting at a given position is
>> strictly bound by the dictionary used, so we can just simply only path
>> through tokens which are right next to each other ignoring all of the rest
>> and *very roughly* should get something close to O(n) runtime. Although
>> I haven't verified this, so maybe.
>>
>> This should be fully backwards compatible if we assume the new algorithm
>> works as I didn't change any public APIs (yet). In another PR I intend to
>> change the APIs a little and expose a Cache struct so that parsing a big
>> batch of text can be done without any allocations after the few first
>> calls. (Which means it's going to be even faster.)
>>
>> By the way, you wrote in the README that we perform significantly worse
>> than mecab. How big is the difference exactly? Do you have a benchmark of
>> this stashed somewhere? Perhaps this is not true anymore? (:
>> ------------------------------
>> You can view, comment on, or merge this pull request online at:
>>
>>   https://github.com/wareya/notmecab-rs/pull/4
>> Commit Summary
>>
>>    - Add two more very simple testcases
>>    - Replace the pathfinding algorithm and run it alongside the old one
>>    - Remove the old pathfinding algorithm
>>
>> File Changes
>>
>>    - *M* Cargo.toml
>>    <https://github.com/wareya/notmecab-rs/pull/4/files#diff-0> (1)
>>    - *M* src/lib.rs
>>    <https://github.com/wareya/notmecab-rs/pull/4/files#diff-1> (349)
>>    - *A* src/pathing.rs
>>    <https://github.com/wareya/notmecab-rs/pull/4/files#diff-2> (408)
>>
>> Patch Links:
>>
>>    - https://github.com/wareya/notmecab-rs/pull/4.patch
>>    - https://github.com/wareya/notmecab-rs/pull/4.diff
>>
>> —
>> You are receiving this because you are subscribed to this thread.
>> Reply to this email directly, view it on GitHub
>> <https://github.com/wareya/notmecab-rs/pull/4?email_source=notifications&email_token=AAEO6EDS2SKTG4353JI2633QGAXQNA5CNFSM4IPCOK42YY3PNVWWK3TUL52HS4DFUVEXG43VMWVGG33NNVSW45C7NFSM4HHDOT7Q>,
>> or mute the thread
>> <https://github.com/notifications/unsubscribe-auth/AAEO6EDMHJ3SZ4BQORG2Z5LQGAXQNANCNFSM4IPCOK4Q>
>> .
>>
>