@koute 2019-08-23
No problem, there is no hurry. (:

Yep, there is a comment on `parse_to_lexertokens` and in the README that the tiebreaking behavior is undefined. To be honest I think that's totally fine. If needed be we technically could always trivially expose an extra API which would return every possible path when there is a tie and just let the user choose.