@koute 2019-08-28
Interesting! So the difference is less than what I thought it could be. I'm... well, I'm not going to say 'pretty sure', but I *think* we can probably beat that (without sacrificing correctness) if we go pedal-to-the-metal. (: